Original adjustment to the grammar was wonky
It only accepted moves like 'N x e4' when, especially for the use case,
moves are expected to not have spaces in them (e.g., Nxe4). As a result
it was put in one terminal rule. This made parsing difficult, so I added
an escape to every move to identify SAN moves (e.g., @Nxe4)