Added my Tests
They mostly work in my runtime eclipse (My checkmate logic raises errors
in my fool's mate games, and that needs to be fixed. Also have chosen
not to implement any scoping functionality, but have developed tests for
them). I will upload the tests as is, and then fix the integration if
time allows.